+++ lol-core/src/core_message.rs
@@ -1,21 +1,17 @@
 use crate::Id;
 
-/// "Req"uest
+/// Request
 #[derive(serde::Serialize, serde::Deserialize, std::fmt::Debug)]
 pub enum Req {
-    InitCluster,
     AddServer(Id),
     RemoveServer(Id),
     ClusterInfo,
     LogInfo,
     HealthCheck,
 }
-/// "Rep"ly
+/// Reply
 #[derive(serde::Serialize, serde::Deserialize, std::fmt::Debug)]
 pub enum Rep {
-    InitCluster {
-        ok: bool,
-    },
     ClusterInfo {
         leader_id: Option<Id>,
         membership: Vec<Id>,