+++ lol-monitor/src/event.rs
@@ -1,6 +1,5 @@
 use std::sync::mpsc;
 use std::thread;
-use std::time::Duration;
 use termion::event::Key;
 use termion::input::TermRead;